More efficient communication with the view server
This change makes the communication between the Erlang VM and
an external view server (couchjs for e.g.) more efficient since
there's now a process which keeps writing to the port (pipe) and
another one which keeps reading from it. This minimizes the
amount of time each endpoint spends blocked on the port read
operation.