Don't call TimeoutFun if no heartbeat was given
If an heartbeat is not specified to the _changes feed, a default
timeout function which returns {stop, Acc} is used. If the changes
filter returns false for a period which reaches the default timeout
(60 secs), the timeout function is called and it causes the seq
tree folding to stop immediately.